move content types to new crate
