+++ examples/cli/serializable.rs
@@ -1,12 +1,8 @@
 use prost::Message;
 use serde::Serialize;
 use valuable::Valuable;
-use xmtp_mls::{
-    codecs::{text::TextCodec, ContentCodec},
-    groups::MlsGroup,
-    storage::group_message::StoredGroupMessage,
-    XmtpApi,
-};
+use xmtp_content_types::{text::TextCodec, ContentCodec};
+use xmtp_mls::{groups::MlsGroup, storage::group_message::StoredGroupMessage, XmtpApi};
 use xmtp_proto::xmtp::mls::message_contents::EncodedContent;
 
 #[derive(Serialize, Debug, Valuable)]
