+++ xmtp_content_types/src/lib.rs
@@ -1,1 +1,28 @@
-// TODO: move content types here
+pub mod group_updated;
+pub mod membership_change;
+#[cfg(test)]
+mod test_utils;
+pub mod text;
+
+use thiserror::Error;
+use xmtp_proto::xmtp::mls::message_contents::{ContentTypeId, EncodedContent};
+
+pub enum ContentType {
+    GroupMembershipChange,
+    GroupUpdated,
+    Text,
+}
+
+#[derive(Debug, Error)]
+pub enum CodecError {
+    #[error("encode error {0}")]
+    Encode(String),
+    #[error("decode error {0}")]
+    Decode(String),
+}
+
+pub trait ContentCodec<T> {
+    fn content_type() -> ContentTypeId;
+    fn encode(content: T) -> Result<EncodedContent, CodecError>;
+    fn decode(content: EncodedContent) -> Result<T, CodecError>;
+}
