+++ xmtp_content_types/src/text.rs
@@ -58,7 +58,7 @@ pub(crate) mod tests {
     #[cfg(target_arch = "wasm32")]
     wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_dedicated_worker);
 
-    use crate::codecs::{text::TextCodec, ContentCodec};
+    use crate::{text::TextCodec, ContentCodec};
 
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test::wasm_bindgen_test)]
     #[cfg_attr(not(target_arch = "wasm32"), test)]
